Refactoring highlightDocument to documentProvider
