+++ src/cloudWatchLogs/commands/addLogEvents.ts
@@ -47,5 +47,4 @@ export async function addLogEvents(
     if (onDidChangeCodeLensEvent) {
         onDidChangeCodeLensEvent.fire()
     }
-    registry.highlightDocument(uri)
 }
